SearchBar component in the /pages directory is now able to query and retrieve the data
pageProps in the _app.tsx will essentially take all of the objects returned via getStaticProps and will send it to the _app.tsx. Thus, when retrieving the parameter value of pageProps, I needed to specify pageProps. Also Components will technically render anything in the immediate directory which is why Index.tsx doesn't need to explicitly be called.