Search bar (#10)
* Incomplete SearchBar functionality, having issues with serverside client calling of the articles props. Unable to obtain the JSON data and unable to iterate through the SearchBar.

* updated env, attempting to establish handler for mongodb querying

* SearchBar component in the /pages directory is now able to query and retrieve the data

pageProps in the _app.tsx will essentially take all of the objects returned via getStaticProps and will send it to the _app.tsx. Thus, when retrieving the parameter value of pageProps, I needed to specify pageProps. Also Components will technically render anything in the immediate directory which is why Index.tsx doesn't need to explicitly be called.

* Searchbar is now able to display based on input text using useEffect, useState Hooks

* Partially styled and formatted Search function

* Possibly create API endpoints in order to reorganize the component

* Created both Change Article and NavBar components to clean up UI, need to style items to match figma

* Added Woof-Logos

* Added the bar between the top-bar and the articles components

Need to make the results of the search not expand the overall higher-header div.
Need to start the dynamic searching to populating into its own [id].

* Styling for the background color and the ChangeArticle component for font and color

* Added to the README.md